recipe_backend/requirements.txt
absl-py==0.14.0
amqp==5.1.1
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==4.1.0
bokeh==2.4.0
branca==0.6.0
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.3
celery==5.2.7
certifi==2024.2.2
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.3.2
chart-studio==1.1.0
clang==5.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.4
colorlover==0.3.0
config==0.5.1
crispy-bootstrap4==2022.1
crispy-bootstrap5==2023.10
criticalpath==0.1.5
cryptography==41.0.1
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.10.2
cufflinks==0.17.3
cycler==0.10.0
dash==2.8.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.13
django-anymail==10.0
django-bootstrap4==23.2
django-crispy-forms==2.0
django-debug-toolbar==3.6.0
django-jazzmin==3.0.0
django-phonenumber-field==7.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
easy-thumbnails==2.8.5
emails==0.6
entrypoints==0.3
filelock==3.12.0
Fiona==1.9.1
Flask==2.2.2
flatbuffers==1.12
flower==1.1.0
folium==0.14.0
fonttools==4.46.0
gast==0.4.0
geopandas==0.12.2
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
graphviz==0.20.1
greenlet==1.1.1
grpcio==1.40.0
gspread==6.1.0
h5py==3.1.0
html5lib==1.1
httplib2==0.22.0
humanize==4.8.0
idna==3.6
imaplib2==3.6
img2table==0.0.19
importlib-metadata==6.0.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==4.3.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kombu==5.3.2
littleutils==0.2.2
lxml==5.2.1
Markdown==3.3.4
MarkupSafe==2.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
munch==2.5.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==3.1
nltk==3.8.1
notebook==6.4.6
num2words==0.5.12
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.5.3.56
opt-einsum==3.3.0
ortools==9.6.2534
outdated==0.2.2
packaging==23.0
pandas==1.5.3
pandas-datareader==0.10.0
pandas-flavor==0.6.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==8.3.2
pingouin==0.5.3
platformdirs==3.5.1
plotly==5.4.0
polars==0.16.14
pooch==1.6.0
pprintpp==0.4.0
premailer==3.10.0
prometheus-client==0.12.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.3
PuLP==2.7.0
pyarrow==11.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycipher==0.5.2
pycparser==2.21
pydot==1.4.2
pydyf==0.8.0
pygame==2.0.1
Pygments==2.10.0
PyInputPlus==0.2.12
PyJWT==2.9.0
PyMuPDF==1.21.1
pyparsing==3.1.2
pyperclip==1.8.2
pyphen==0.14.0
pyproj==3.4.1
pyrsistent==0.18.0
PySimpleValidate==0.2.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==302
pywinpty==1.1.6
PyYAML==5.4.1
pyzmq==22.3.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.3
rsa==4.9
scikit-learn==1.0
scipy==1.11.1
seaborn==0.11.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.25
sqlparse==0.4.4
statsmodels==0.14.0
stdiomask==0.0.6
StrEnum==0.4.15
stripe==4.0.2
sympy==1.9
tabulate==0.9.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tinycss2==1.2.1
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.6.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wrapt==1.12.1
xarray==2023.2.0
xlrd==2.0.1
XlsxWriter==3.0.9
zipp==3.11.0
zopfli==0.2.3
